Use generated cargo-sources instead of tagged vendor sources
This'll allow us to ensure that the project builds properly on Flathub
before making a tagged release, since currently building on Flathub
requires there to be a vendor source which gets created on a tagged
commit and if something breaks with that release, there is a need to
make another minor release to fix the build issue. This is not ideal,
hence the cargo-sources.